pyproject.toml
[tool.poetry]
name = "ruleset-conv"
version = "0.1.0"
description = "将classical规则集拆分为domain与ipcidr规则集。"
authors = ["Yorun <yorun@yorun.me>"]

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^6.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
